run: full mechanism battery re-run, 11 families, 3,800 datasets
One command, attempt 1, first run after the last code change, exit 0 in 6,354.4 s
on four CPU workers. Eleven families as 15 conditions: the original 11 conditions
at 100 datasets at each of n = 350 and n = 3,600, plus the four element 12
part (iv) families at 400 datasets at n = 350, each fitted at rho = 0 with 200
bootstrap refits, plus a 20-dataset PyMC subset per family.

The three required outcomes of A2 element 11(a) still PASS. Cross-checks: the
vectorised rho curve against sensitivity_sweep, max abs NIE difference 0.00335
over 198 comparisons; rho*_point against breakdown_frontier, max abs difference
0.00096 over 33 datasets; each family's Monte Carlo truth against its analytic
truth, max abs difference 0.000773 over 15 conditions at 2,000,000 rows.

THE MAXIMUM-LIKELIHOOD PATH DID NOT MOVE. All 22 shared condition-and-size cells
reproduce the 2026-09-07 run exactly, same coverage count out of the same
denominator and same bias to four decimals. The prior change is confined to the
posterior path by construction and this is the check, not the claim; it also
clears the one refactor that could have moved a number (the bootstrap now
resamples len(x) rows rather than the requested n).

The PyMC subset reproduces gate (b) to the dataset on all seven original
families: f4 7/20 -> 20/20 and f7 12/20 -> 19/20, the other five unchanged, max
r_hat 1.0000 and zero divergences over all 220 fits.

What the four new families found (400 datasets each at n = 350):
- f9_varying_variance is the expensive one: NIE bias +0.0653 +/- 0.0015 and
  coverage 169/400 [0.374, 0.473]. The bias is entirely on the mediated path
  (NDE bias -0.0018, coverage 375/400). A cue that widens the spread of the
  reasoning as well as shifting its level is read by a single-sigma_m model as a
  larger mediated effect.
- f10_sparse_groups is an interval problem and not a bias problem: bias -0.0009
  +/- 0.0018 with coverage 362/400 [0.872, 0.932], because the row bootstrap
  resamples rows that are not independent. A per-item table cannot use one.
- f11_mediator_missingness keeps 264.5 of 350 rows, drops the long traces, and
  costs bias -0.0108 with coverage 367/400. f8_nonlinear_depth costs -0.0156
  with coverage 359/400.

Two linter nits were fixed after the run (int(math.ceil(x)) to math.ceil(x) in
the family 10 generator, .items() to .values() in one test loop). Neither can
change a number and that is checked rather than asserted: redrawing family 10's
dataset 0 after the edit reproduces the run's own dataset_rows.csv row exactly
(answer rate 0.702857142857, mediator mean 6.652451383811, corr(M, Y)
0.592962464443). tests/test_mechanism_battery.py 35/35 after the edit.

Full write-up with every denominator: docs/ESTIMATOR-PRIORS-2026-09-07.md
sections 4 to 6.

+++ experiments/mechanism_battery.py
@@ -458,7 +458,7 @@ class SparseGroups(Mechanism):
         self.analytic_truth = (0.0, p1 - p0, p1 - p0)
 
     def noise(self, rng, n):
-        n_groups = int(math.ceil(n / GROUP_SIZE))
+        n_groups = math.ceil(n / GROUP_SIZE)
         per_group = rng.normal(0.0, GROUP_SD, n_groups)
         group_of_row = np.arange(n) // GROUP_SIZE
         return {
